requirements.txt
Django~=5.1.1
djangorestframework~=3.15.2
openai~=1.44.1
requests~=2.32.3
cryptography~=43.0.1
django-cors-headers
djangorestframework-simplejwt